+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Report elements that have cursor pointer style.
 
+### Changed
+- Only report ARIA identification attributes when they uniquely identify the element.
+
 ## 0.1.8 - 2025-12-12
 
 ### Changed